Add map object for item images
Image references to local assets in React Native are statically compiled and not resolved at runtime, so there's no way to do a dynamic `require()` call. Because we want at least some image assets stored locally, we have to manually assemble an object that maps properties to `require()` calls, which can then be used to dynamically pick the image.